Use find: to find all packages
When listing the packages manually we would need to list also all
subpackages which is hard to keep up to date. We can use find: to find
packages automatically. For this to work we need to use __init__.py
files in all the packages. We could use namespace packages but we would
run into other issues most likely eventually, and as we don't need the
feature of namespace packages, let's use normal packages with init files
instead.