Introduce work log to resume run
Introduce a work log that logs everytime we do an incremental update so
we can resume. With the worklog we can see which files we have already
processed into which intermediate files. We can then skip all the files
that were already successfully processed, and rerun on all the one that
were not succesfully processed.

Afterwards we can watch for new files normally