Remove traceback from nomadic start error
Creating a workspace with the same name
as one that already exists previously threw
a traceback, which I think is not necessary.

I removed it in favour of a simpler click error
message.

I also made a small refactor to link workspace initialisation
to the class, while still allowing fetching an object without
producing the directories.