Fix truth value of series is ambigeous error
The error was caused when still a website was opened with a different
panel than the current running one. This way filtering by target region
can return an empty series and the code later will error.

Fixing it by just not executing following code, as this is not really a
supported case.