Allow fastqdir path to not exist yet
We want to allow the user to start nomadic directly and some of the
output directories might not exist yet.

We check that the path that the
user provides exists, or error. But the user does can also only provide
the path only of the data directory of minknow. In this case we will
keep checking for the folders with a glob pattern and only warn the
users that the files don't exist yet.