Retain unfiltered VCF file
In the future we will want to merge variant
calling results for multiple experiments.
However, some mutations may be present in only
one experiment. If we filter our VCF file to only
contain variant positions; no information about
this position may exist for the experiments
where it is not variable. As a result, we
do not know whether the position failed QC,
or was homozygous for all samples. Both
lead to the variant not be represented.
Hence we must retain the unfiltered VCF.