Make sample_type mandatory
Make sample_type mandatory. We need it for summary analysis and it is
best if people just include it when creating the sample sheet. It should
not be much extra work.

We maybe want to make it optional if we can derive it from the sample
name, to be discussed.