Do not seperate mutations by alt alleles
For now just don't group them by alt alleles. We might have a different
set of alt alleles when we call the same mutation in different
experiments, as we might have a triallelic side. As we in the end group
by amino acid change, this leads to problems. We need to better discuss
how to hanle csq calling correctly for multiple changes.