Allow to only view with nomadic summarize command
We try it first like this, but imo it would be better to actually
separate viewing the summary into a different command or combine it with
dashboard. It is a bit awkward to combine an output command (produce the
files of summary) with an input command (showing files from a folder).